requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
requests==2.31.0
pillow==10.1.0
google-generativeai==0.3.2
pymupdf==1.23.8
speechrecognition==3.10.0
